+++ packages/elements/src/components/ui/cards/CardBody.tsx
@@ -1,0 +1,8 @@
+import { BoxProps, Column } from "@stenajs-webui/core";
+import * as React from "react";
+
+interface CardBodyProps extends Omit<BoxProps, "indent" | "spacing"> {}
+
+export const CardBody: React.FC<CardBodyProps> = (props) => (
+  <Column indent={2} spacing={2} {...props} />
+);